+++ scripts/desktop_bannedip.sh
@@ -8,9 +8,9 @@
 
 source /etc/lsb-release
 
-if [ "x$2" == "x" ]; then
-   echo "***** Edit of crontab on remote servers *****"
-   echo "Usage:   $0  hostfile  [hostgrouporname]"
+if [ "x$3" == "x" ]; then
+   echo "***** Get/Edit banned IP on a remote servers *****"
+   echo "Usage:   $0  hostfile  [hostgrouporname] list|unban  IP"
    echo "         [hostgrouporname] can be 'master', 'deployment', 'web', 'remotebackup', or list separated with comma like 'master,deployment' (default)"
    echo "Example: $0  myhostfile  master,deployment"
    echo "Example: $0  myhostfile  withX.mysellyoursaasdomain.com"
@@ -22,10 +22,9 @@ if [ "x$target" == "x" ]; then
 	target="master"
 fi
 
-username=$3
-if [ "x$username" == "x" ]; then
-	username="root"
-fi
+action=$3
+
+ip=$4
 
 export currentpath=$(dirname "$0")
 
@@ -34,8 +33,12 @@ cd $currentpath/ansible
 echo "Execute ansible for host group $1 and targets $target"
 pwd
 
-
-command='ansible-playbook -K change_crontab_server.yml -i hosts-'$1' -e "target='$target' username='$username'"' 
+if [ "x$action" == "xlist" ]; then
+	command='ansible-playbook -K bannedip_list.yml -i hosts-'$1' -e "target='$target'"' 
+fi
+if [ "x$action" == "xunban" ]; then
+	command='ansible-playbook -K bannedip_unban.yml -i hosts-'$1' -e "target='$target' ip='$ip'"' 
+fi
 
 echo "$command"
 eval $command